Changes to read .tsv files in Omic files
Because

--TSV are superior to CSV
--Data operations has shifted to using tsv, the liftover and annotation
--pipeline use these preferentially

Changed the omic loader as minimally as possible to incorporate tsv files. Changes including checking for data.tsv or $modelId.tsv in the LoadOmic class. Changes in the UtilityService include adding a delimiter parameter in serializeCSVToMaps allowing the change between commas and tabs. Previous calls to this method were changed to include the second argument as ","