Created a more robust approach to linking Raw Data Urls
Because
*Multiple accession types are available for linking to different resources
*The template contain heaps and heap of rubbish in this column that needs filtering

Created the formatAccesionsToURL function in the DomainObjectCreator. This function will proccess a select few accession with the ENA, EGA, and GEO returning a built URL and the original accession. If not, it will return and empty string to avoid filling the pdxFinder with rubbish. An additional variable on MolecularDataEntryDTO was created to store the link and is then retrieved by Thyme Leaf to generated the front end link. An empty string "" in the raw data url will cause a "Not Available to show up on the front end"