Created behavior to validate columns by regex
Because:
- Invalid characters need to be detected.
- Need for custom regex restrictions

Added the three new classes. IllegalValueError and IllegalValueErrorCreator following the patterns as the rest of the validator.
ValueRestrictions is a class to be expanded upon in future iteration to allow for a list of values to restrict. Now it is
used to hold both the regex and the description for error printing. Only restriction currently are Id's are to be URL_SAFE

Revert "Created behavior to validate columns by regex" to remove unintented changes

This reverts commit 6532875c