Added categories for all relevant columns in all sheets of metadata.
Refactoring ValueRestrictions to use Predicates to be more robust. Clarified predicate logic in ValueRestrictions class.
Added canBeEmpty to value restriction. This will be set so that columns can ignore the presence of blanks if neccessary.
Working on changes to detect empty column in the EmptyValueErrorCreator. Reorganized directory structure. Added new class
to hold ValueRestrictions definitions to declutter the PdxValidationRuleset. Refactore PdxValidationRuleset to remove
boilplate code.